Updated 1)DAOs 2)Entities 3)Content data base 4)Fragments about 5)App class
Added fragments and view models:
1)catalog, tabs(my and all routes), splash, settings, favorite
2)objects, route details
3)objects details, orders

Added unstable navigation graphs

+++ app/src/main/java/com/grsu/guideapp/database/content/entities/ListPoi.kt
@@ -28,5 +28,5 @@ class ListPoi(
     @ColumnInfo(name = "id_route", index = true) val id_route: Long,
     @ColumnInfo(name = "id_poi", index = true) val id_poi: Long,
     @ColumnInfo(name = "id_point") val id_point: String,
-    @ColumnInfo(name = "distance") val distance: Int
+    @ColumnInfo(name = "distance") val distance: Double
 )
